Got why distance between seeds looked wrong:
The covariance matrix from their mutations wasn't representative of
the variations between each other. That's quite interesting because it
means the input generated are "far" from the seeds themselves.
Regardless, we'd like to a have a "global" basis that works reasonnably
well for any test case/seed. Hopefully it will get better once we
activate the evolutionnary algorithm.

Next TODO: variances are wrongs.