Mistage in scheduling and divergence computation
Divergence: forgot to divide the covariance matrices by number of
samples. So the computed determinant were wrong.

Also, the scheduler had a problem when a seed was discovered "close to
the end" of the campaign. When there wasn't enough threads to use all
seeds, the surplus of thread was just "forgotten". Hence, in the even
that new seeds where discovered, which could leverage these threads,
since they were forgotten, the scheduler couldn't use them.
To solve this problem, use a separate list to store them (and not forget
them). Reactive these threads when new seed appear.